Rename: 自分の所属チーム情報をplayer.moduleで管理
